requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
deepface==0.0.93
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flatbuffers==24.3.25
future==1.0.0
gast==0.6.0
gdown==5.2.0
geocoder==1.38.1
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-cloud-texttospeech==2.21.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googlemaps==4.10.0
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
MarkupSafe==2.1.5
ml_dtypes==0.4.1
mtcnn==1.0.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.11.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
tf-keras
